refactor: create notifications list
